refactor deindexing code into a new module
